告警修改

+++ rosen/modules/texgine/src/typography_impl.cpp
@@ -425,7 +425,7 @@ void TypographyImpl::ComputeRoundRect(VariantSpan& span, int& index, int& preInd
     bool leftRound = false;
     bool rightRound = false;
     if (span.HasBackgroundRect()) {
-        int lineSpanCount = metric.lineSpans.size();
+        int lineSpanCount = static_cast<int>(metric.lineSpans.size());
         int styleId = span.GetTextStyle().styleId;
         // index - 1 is previous index, -1 is the invalid styleId
         int preStyleId = index == 0 ? -1 : metric.lineSpans[index - 1].GetTextStyle().styleId;
